@@ -1,101 +1,95 @@
 <Project Sdk="Microsoft.NET.Sdk.Web">
 
-  <PropertyGroup>
-    <TargetFramework>net5.0</TargetFramework>
-    <TypeScriptCompileBlocked>true</TypeScriptCompileBlocked>
-    <TypeScriptToolsVersion>Latest</TypeScriptToolsVersion>
-    <IsPackable>false</IsPackable>
-    <SpaRoot>ClientApp\</SpaRoot>
-    <DefaultItemExcludes>$(DefaultItemExcludes);$(SpaRoot)node_modules\**</DefaultItemExcludes>
+	<PropertyGroup>
+		<TargetFramework>net7.0</TargetFramework>
+		<Nullable>enable</Nullable>
+		<IsPackable>false</IsPackable>
+		<SpaRoot>ClientApp\</SpaRoot>
+		<SpaProxyServerUrl>https://localhost:4200</SpaProxyServerUrl>
+		<SpaProxyLaunchCommand>npm start</SpaProxyLaunchCommand>
+		<ImplicitUsings>enable</ImplicitUsings>
+	</PropertyGroup>
 
-    <!-- Set this to true if you enable server-side prerendering -->
-    <BuildServerSideRenderer>false</BuildServerSideRenderer>
-  </PropertyGroup>
+	<ItemGroup>
+		<PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="12.0.0" />
+		<PackageReference Include="Microsoft.AspNetCore.SpaServices.Extensions" Version="7.0.0" />
+		<PackageReference Include="Microsoft.AspNetCore.SpaProxy" Version="7.0.0" />
+		<PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="7.0.0" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <PackageReference Include="AutoMapper.Extensions.Microsoft.DependencyInjection" Version="8.1.1" />
-    <PackageReference Include="Microsoft.AspNetCore.SpaServices.Extensions" Version="5.0.3" />
-    <PackageReference Include="Microsoft.VisualStudio.Web.CodeGeneration.Design" Version="5.0.2" />
-    <PackageReference Include="Serilog.AspNetCore" Version="3.4.0" />
-    <PackageReference Include="Serilog.Extensions.Logging.File" Version="2.0.0" />
-    <PackageReference Include="Swashbuckle.AspNetCore" Version="6.0.5" />
-  </ItemGroup>
+	<ItemGroup>
+		<!-- Don't publish the SPA source files, but do show them in the project files list -->
+		<Content Remove="$(SpaRoot)**" />
+		<None Remove="$(SpaRoot)**" />
+		<None Include="$(SpaRoot)**" Exclude="$(SpaRoot)node_modules\**" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <!-- Don't publish the SPA source files, but do show them in the project files list -->
-    <Content Remove="$(SpaRoot)**" />
-    <None Remove="$(SpaRoot)**" />
-    <None Include="$(SpaRoot)**" Exclude="$(SpaRoot)node_modules\**" />
-  </ItemGroup>
+	<ItemGroup>
+		<None Remove="ClientApp\src\app\app.routing.module.ts" />
+		<None Remove="ClientApp\src\app\core\alertdialog\alertdialog.component.ts" />
+		<None Remove="ClientApp\src\app\core\http\encounter.service.ts" />
+		<None Remove="ClientApp\src\app\core\http\http.option.ts" />
+		<None Remove="ClientApp\src\app\core\interceptors\server.error.interceptor.ts" />
+		<None Remove="ClientApp\src\app\modules\encounter\encounter.detail.component.ts" />
+		<None Remove="ClientApp\src\app\modules\encounter\encounter.list.component.ts" />
+		<None Remove="ClientApp\src\app\modules\encounter\encounter.list.service.ts" />
+		<None Remove="ClientApp\src\app\shared\components\app.language.menu.component.ts" />
+		<None Remove="ClientApp\src\app\shared\constants\constant.ts" />
+		<None Remove="ClientApp\src\app\shared\directive\sortable.directive.ts" />
+		<None Remove="ClientApp\src\app\shared\models\encounter.model.ts" />
+		<None Remove="ClientApp\src\app\shared\pipes\savingthrow.pipe.ts" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <None Remove="ClientApp\src\app\app.routing.module.ts" />
-    <None Remove="ClientApp\src\app\core\alertdialog\alertdialog.component.ts" />
-    <None Remove="ClientApp\src\app\core\http\encounter.service.ts" />
-    <None Remove="ClientApp\src\app\core\http\http.option.ts" />
-    <None Remove="ClientApp\src\app\core\interceptors\server.error.interceptor.ts" />
-    <None Remove="ClientApp\src\app\modules\encounter\encounter.detail.component.ts" />
-    <None Remove="ClientApp\src\app\modules\encounter\encounter.list.component.ts" />
-    <None Remove="ClientApp\src\app\modules\encounter\encounter.list.service.ts" />
-    <None Remove="ClientApp\src\app\shared\components\app.language.menu.component.ts" />
-    <None Remove="ClientApp\src\app\shared\constants\constant.ts" />
-    <None Remove="ClientApp\src\app\shared\directive\sortable.directive.ts" />
-    <None Remove="ClientApp\src\app\shared\models\encounter.model.ts" />
-    <None Remove="ClientApp\src\app\shared\pipes\savingthrow.pipe.ts" />
-  </ItemGroup>
+	<ItemGroup>
+		<ProjectReference Include="..\REG.Core\REG.Core.csproj" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <ProjectReference Include="..\REG.Core\REG.Core.csproj" />
-  </ItemGroup>
+	<ItemGroup>
+		<TypeScriptCompile Include="ClientApp\src\app\app.routing.module.ts" />
+		<TypeScriptCompile Include="ClientApp\src\app\core\alertdialog\alertdialog.component.ts" />
+		<TypeScriptCompile Include="ClientApp\src\app\core\http\encounter.service.ts" />
+		<TypeScriptCompile Include="ClientApp\src\app\core\http\http.option.ts" />
+		<TypeScriptCompile Include="ClientApp\src\app\core\interceptors\server.error.interceptor.ts" />
+		<TypeScriptCompile Include="ClientApp\src\app\modules\encounter\encounter.detail.component.ts" />
+		<TypeScriptCompile Include="ClientApp\src\app\modules\encounter\encounter.list.component.ts" />
+		<TypeScriptCompile Include="ClientApp\src\app\modules\encounter\encounter.list.service.ts" />
+		<TypeScriptCompile Include="ClientApp\src\app\shared\components\app.language.menu.component.ts" />
+		<TypeScriptCompile Include="ClientApp\src\app\shared\constants\constant.ts" />
+		<TypeScriptCompile Include="ClientApp\src\app\shared\directive\sortable.directive.ts" />
+		<TypeScriptCompile Include="ClientApp\src\app\shared\models\encounter.model.ts" />
+		<TypeScriptCompile Include="ClientApp\src\app\shared\pipes\savingthrow.pipe.ts" />
+	</ItemGroup>
 
-  <ItemGroup>
-    <TypeScriptCompile Include="ClientApp\src\app\app.routing.module.ts" />
-    <TypeScriptCompile Include="ClientApp\src\app\core\alertdialog\alertdialog.component.ts" />
-    <TypeScriptCompile Include="ClientApp\src\app\core\http\encounter.service.ts" />
-    <TypeScriptCompile Include="ClientApp\src\app\core\http\http.option.ts" />
-    <TypeScriptCompile Include="ClientApp\src\app\core\interceptors\server.error.interceptor.ts" />
-    <TypeScriptCompile Include="ClientApp\src\app\modules\encounter\encounter.detail.component.ts" />
-    <TypeScriptCompile Include="ClientApp\src\app\modules\encounter\encounter.list.component.ts" />
-    <TypeScriptCompile Include="ClientApp\src\app\modules\encounter\encounter.list.service.ts" />
-    <TypeScriptCompile Include="ClientApp\src\app\shared\components\app.language.menu.component.ts" />
-    <TypeScriptCompile Include="ClientApp\src\app\shared\constants\constant.ts" />
-    <TypeScriptCompile Include="ClientApp\src\app\shared\directive\sortable.directive.ts" />
-    <TypeScriptCompile Include="ClientApp\src\app\shared\models\encounter.model.ts" />
-    <TypeScriptCompile Include="ClientApp\src\app\shared\pipes\savingthrow.pipe.ts" />
-  </ItemGroup>
+	<ItemGroup>
+		<None Update="ClientApp\src\assets\i18n\en.json">
+			<CopyToOutputDirectory>Never</CopyToOutputDirectory>
+		</None>
+	</ItemGroup>
 
-  <ItemGroup>
-    <None Update="ClientApp\src\assets\i18n\en.json">
-      <CopyToOutputDirectory>Never</CopyToOutputDirectory>
-    </None>
-  </ItemGroup>
+	<Target Name="DebugEnsureNodeEnv" BeforeTargets="Build" Condition=" '$(Configuration)' == 'Debug' And !Exists('$(SpaRoot)node_modules') ">
+		<!-- Ensure Node.js is installed -->
+		<Exec Command="node --version" ContinueOnError="true">
+			<Output TaskParameter="ExitCode" PropertyName="ErrorCode" />
+		</Exec>
+		<Error Condition="'$(ErrorCode)' != '0'" Text="Node.js is required to build and run this project. To continue, please install Node.js from https://nodejs.org/, and then restart your command prompt or IDE." />
+		<Message Importance="high" Text="Restoring dependencies using 'npm'. This may take several minutes..." />
+		<Exec WorkingDirectory="$(SpaRoot)" Command="npm install" />
+	</Target>
 
-  <Target Name="DebugEnsureNodeEnv" BeforeTargets="Build" Condition=" '$(Configuration)' == 'Debug' And !Exists('$(SpaRoot)node_modules') ">
-    <!-- Ensure Node.js is installed -->
-    <Exec Command="node --version" ContinueOnError="true">
-      <Output TaskParameter="ExitCode" PropertyName="ErrorCode" />
-    </Exec>
-    <Error Condition="'$(ErrorCode)' != '0'" Text="Node.js is required to build and run this project. To continue, please install Node.js from https://nodejs.org/, and then restart your command prompt or IDE." />
-    <Message Importance="high" Text="Restoring dependencies using 'npm'. This may take several minutes..." />
-    <Exec WorkingDirectory="$(SpaRoot)" Command="npm install" />
-  </Target>
+	<Target Name="PublishRunWebpack" AfterTargets="ComputeFilesToPublish">
+		<!-- As part of publishing, ensure the JS resources are freshly built in production mode -->
+		<Exec WorkingDirectory="$(SpaRoot)" Command="npm install" />
+		<Exec WorkingDirectory="$(SpaRoot)" Command="npm run build -- --configuration production" />
 
-  <Target Name="PublishRunWebpack" AfterTargets="ComputeFilesToPublish">
-    <!-- As part of publishing, ensure the JS resources are freshly built in production mode -->
-    <Exec WorkingDirectory="$(SpaRoot)" Command="npm install" />
-    <Exec WorkingDirectory="$(SpaRoot)" Command="npm run build -- --prod" />
-    <Exec WorkingDirectory="$(SpaRoot)" Command="npm run build:ssr -- --prod" Condition=" '$(BuildServerSideRenderer)' == 'true' " />
-
-    <!-- Include the newly-built files in the publish output -->
-    <ItemGroup>
-      <DistFiles Include="$(SpaRoot)dist\**; $(SpaRoot)dist-server\**" />
-      <DistFiles Include="$(SpaRoot)node_modules\**" Condition="'$(BuildServerSideRenderer)' == 'true'" />
-      <ResolvedFileToPublish Include="@(DistFiles->'%(FullPath)')" Exclude="@(ResolvedFileToPublish)">
-        <RelativePath>%(DistFiles.Identity)</RelativePath>
-        <CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
-        <ExcludeFromSingleFile>true</ExcludeFromSingleFile>
-      </ResolvedFileToPublish>
-    </ItemGroup>
-  </Target>
+		<!-- Include the newly-built files in the publish output -->
+		<ItemGroup>
+			<DistFiles Include="$(SpaRoot)dist\**; $(SpaRoot)dist-server\**" />
+			<ResolvedFileToPublish Include="@(DistFiles->'%(FullPath)')" Exclude="@(ResolvedFileToPublish)">
+				<RelativePath>wwwroot\%(RecursiveDir)%(FileName)%(Extension)</RelativePath>
+				<CopyToPublishDirectory>PreserveNewest</CopyToPublishDirectory>
+				<ExcludeFromSingleFile>true</ExcludeFromSingleFile>
+			</ResolvedFileToPublish>
+		</ItemGroup>
+	</Target>
 
 </Project>
