@@ -1,10 +1,9 @@
-namespace REG.Core.Abstractions.Services.Models.Json
+namespace REG.Core.Abstractions.Services.Models.Json;
+
+public class LegendaryAction
 {
-    public class LegendaryAction
-    {
-        public string Name { get; set; }
-        public string Desc { get; set; }
-        public int Attack_Bonus { get; set; }
-        public string Damage_Dice { get; set; }
-    }
+    public string Name { get; set; }
+    public string Desc { get; set; }
+    public int Attack_Bonus { get; set; }
+    public string Damage_Dice { get; set; }
 }
