@@ -1,9 +1,8 @@
-namespace REG.Core.Abstractions.Services.Models.Json
+namespace REG.Core.Abstractions.Services.Models.Json;
+
+public class Reaction
 {
-    public class Reaction
-    {
-        public string Name { get; set; }
-        public string Desc { get; set; }
-        public int Attack_Bonus { get; set; }
-    }
+    public string Name { get; set; }
+    public string Desc { get; set; }
+    public int Attack_Bonus { get; set; }
 }
