@@ -1,28 +1,27 @@
-﻿namespace REG.Core.Domain
+﻿namespace REG.Core.Domain;
+
+public enum Difficulty
 {
-    public enum Difficulty
-    {
-        Easy = 0,
-        Medium,
-        Hard,
-        Deadly
-    }
-    public enum MonsterType
-    {
-        Aberration = 1,
-        Beast,
-        Celestial,
-        Construct,
-        Dragon,
-        Elemental,
-        Fey,
-        Fiend,
-        Giant,
-        Humanoid,
-        Monstrosity,
-        Ooze,
-        Plant,
-        SwarmOfTinyBeasts,
-        Undead
-    }
+    Easy = 0,
+    Medium,
+    Hard,
+    Deadly
+}
+public enum MonsterType
+{
+    Aberration = 1,
+    Beast,
+    Celestial,
+    Construct,
+    Dragon,
+    Elemental,
+    Fey,
+    Fiend,
+    Giant,
+    Humanoid,
+    Monstrosity,
+    Ooze,
+    Plant,
+    SwarmOfTinyBeasts,
+    Undead
 }
