@@ -10,7 +10,7 @@ import { EncounterService } from '../../core/http/encounter.service';
     providers: [EncounterService]
   })
 export class EncounterDetailComponent {
-  @Input() detail: EncounterDetailModel;
+  @Input() detail!: EncounterDetailModel;
 
   constructor(public activeModal: NgbActiveModal) {
   }
