@@ -1,6 +1,6 @@
 import { Component } from '@angular/core';
 import { environment } from '../../../environments/environment';
-import { version } from '../../../../package.json';
+const { version: version } = require('../../../../package.json')
 
 @Component({
     selector: 'app-footer',
