@@ -1,72 +1,72 @@
 export class SpecialAbility {
-  name: string;
-  desc: string;
-  attack_Bonus: number;
-  damage_Dice: string;
+  name: string = "";
+  desc: string = "";
+  attack_Bonus: number = 0;
+  damage_Dice: string = "";
 }
 export class Action {
-  name: string;
-  desc: string;
-  attack_Bonus: number;
-  damage_Dice: string;
-  damage_Bonus: number;
+  name: string = "";
+  desc: string = "";
+  attack_Bonus: number = 0;
+  damage_Dice: string = "";
+  damage_Bonus: number = 0;
 }
 export class LegendaryAction {
-  name: string;
-  desc: string;
-  attack_Bonus: number;
-  damage_Dice: string;
+  name: string = "";
+  desc: string = "";
+  attack_Bonus: number = 0;
+  damage_Dice: string = "";
 }
 export class Reaction {
-  name: string;
-  desc: string;
-  attack_Bonus: number;
+  name: string = "";
+  desc: string = "";
+  attack_Bonus: number = 0;
 }
 export class EncounterDetailModel {
-  xp: number;
-  count: number;
-  name: string;
-  type: string;
-  difficulty: string;
-  challengeRating: string;
-  size: string;
-  alignment: string;
-  hp: number;
-  ac: number;
-  hitDice: string;
-  speed: string;
-  senses: string;
-  languages: string;
-  strength: number;
-  dexterity: number;
-  constitution: number;
-  intelligence: number;
-  wisdom: number;
-  charisma: number;
-  strengthSave: number;
-  constitutionSave: number;
-  dexteritySave: number;
-  intelligenceSave: number;
-  wisdomSave: number;
-  charismaSave: number;
-  history: number;
-  perception: number;
-  damageVulnerabilities: string;
-  damageResistances: string;
-  damageImmunities: string;
-  conditionImmunities: string;
-  specialAbilities: SpecialAbility[];
-  actions: Action[];
-  reactions: Reaction[];
-  legendaryActions: LegendaryAction[];
+  xp: number = 0;
+  count: number = 0;
+  name: string = "";
+  type: string = "";
+  difficulty: string = "";
+  challengeRating: string = "";
+  size: string = "";
+  alignment: string = "";
+  hp: number = 0;
+  ac: number = 0;
+  hitDice: string = "";
+  speed: string = "";
+  senses: string = "";
+  languages: string = "";
+  strength: number = 0;
+  dexterity: number = 0;
+  constitution: number = 0;
+  intelligence: number = 0;
+  wisdom: number = 0;
+  charisma: number = 0;
+  strengthSave: number = 0;
+  constitutionSave: number = 0;
+  dexteritySave: number = 0;
+  intelligenceSave: number = 0;
+  wisdomSave: number = 0;
+  charismaSave: number = 0;
+  history: number = 0;
+  perception: number = 0;
+  damageVulnerabilities: string = "";
+  damageResistances: string = "";
+  damageImmunities: string = "";
+  conditionImmunities: string = "";
+  specialAbilities: SpecialAbility[] = [];
+  actions: Action[] = [];
+  reactions: Reaction[] = [];
+  legendaryActions: LegendaryAction[] = [];
 }
 export class EncounterModel {
-  sumXp: number;
-  encounters: EncounterDetailModel[];
+  sumXp: number = 0;
+  encounters: EncounterDetailModel[] = [];
 }
 export class EncounterOptionModel {
-  partyLevel: number;
-  partySize: number;
+  partyLevel: number = 0;
+  partySize: number = 0;
   difficulty: number | null;
   monsterTypes: number[];
   count = 10;
